feat: treasury vault — оптимизация контракта + backend + frontend
Контракт (239 KB, -17%):
- opt-level=z, lto=fat, strip=true, panic=abort
- anchor-spl minimal features (no defaults)
- deposit_count/payout_count: u64 → u32
- Убраны msg!() вызовы (events достаточно)
- 16/16 тестов проходят

Backend: TreasuryModule с deposit/payout/getVaultInfo + cron
Frontend: TreasuryInfo компонент + store + Cash page интеграция

+++ contracts/treasury-vault/programs/treasury-vault/src/events.rs
@@ -14,7 +14,7 @@ pub struct DepositEvent {
     pub vault: Pubkey,
     pub amount: u64,
     pub total_deposited: u64,
-    pub deposit_count: u64,
+    pub deposit_count: u32,
     pub timestamp: i64,
 }
 
@@ -24,7 +24,7 @@ pub struct PayoutEvent {
     pub payout_wallet: Pubkey,
     pub amount: u64,
     pub total_paid_out: u64,
-    pub payout_count: u64,
+    pub payout_count: u32,
     pub timestamp: i64,
 }
